Moved components around
