+++ client/src/components/admin/panels/section/update-card-section/update-card-inventory.tsx
