+++ client/src/providers/card-provider.ts
@@ -163,6 +163,8 @@ class CardProvider {
     }
   };
 
+  public getCardsFromSetFromAPI = async (setName: string) => {};
+
   public getAPIID = async (
     setName: string,
     cardName: string
